requirements.txt
python-telegram-bot==13.15
environs==11.0.0
urllib3==2.2.1
